Final changes for deploment instruction

+++ includes/class-loyalty-discount.php
@@ -1,13 +1,13 @@
 <?php
 class loyaltydiscount{
   public array $prodDiscount;
-  public int $t;
+  public float $t;
   public $tax_toggle;
   public function __construct(array $prodDiscount, int $t)
   {
     $this->prodDiscount = $prodDiscount;
     $this->t = $t;
-  
+
   }
 
 }
